mrp2 2. test öncesi maddeler yapıldı

+++ TsiErp.Entities/Entities/PlanningManagement/MRPLine/Dtos/CreateMRPLinesDto.cs
@@ -19,7 +19,7 @@ namespace TsiErp.Entities.Entities.PlanningManagement.MRPLine.Dtos
         /// <summary>
         /// Miktar
         /// </summary>
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         /// <summary>
         /// Stoktan Kullanılacak
         /// </summary>
@@ -43,7 +43,7 @@ namespace TsiErp.Entities.Entities.PlanningManagement.MRPLine.Dtos
         /// <summary>
         /// İhtiyaç Miktar
         /// </summary>
-        public int RequirementAmount { get; set; }
+        public decimal RequirementAmount { get; set; }
         /// <summary>
         /// Hesaplanacak
         /// </summary>
